+++ src/components/NavMenu/MenuCategory/Device/Device.scss
@@ -7,7 +7,4 @@
   &:hover {
     font-weight: bold;
   }
-  &:hover::after {
-    position: absolute;
-  }
 }
