+++ src/components/NavMenu/MenuCategory/MenuCategory.scss
@@ -21,7 +21,7 @@
     position: absolute;
     top: 45px;
     left: 450px;
-    width: 300px;
+    width: 100%;
     animation: slideMenu 1s;
   }
 
@@ -33,10 +33,23 @@
     border-left: 1px solid color.$input-color;
     animation: slideImg 1s;
 
-    img {
+    > div {
       width: 400px;
       height: 400px;
       border-radius: 20px;
+      overflow: hidden;
+
+      img {
+        width: 400px;
+        height: 400px;
+        border-radius: 20px;
+        transform: scale(1);
+
+        &:hover {
+          transform: scale(1.2);
+          transition-duration: 1s;
+        }
+      }
     }
   }
 }
