--- src/components/NavMenu/UserMenu.js
@@ -1,8 +1,0 @@
-import React from "react";
-import "./UserMenu.scss";
-
-const UserMenu = ({ menu }) => {
-  return <li className="nav-icon-menu">{menu}</li>;
-};
-
-export default UserMenu;
