+++ src/components/UserMenu/UserMenuItem/UserMenuItem.js
@@ -1,0 +1,8 @@
+import React from "react";
+import "./UserMenuItem.scss";
+
+const UserMenuItem = ({ menu }) => {
+  return <li className="nav-icon-menu">{menu}</li>;
+};
+
+export default UserMenuItem;
