+++ src/components/UserMenu/UserMenuItem/UserMenuItem.scss
@@ -1,6 +1,7 @@
 .nav-icon-menu {
   margin: 5px 0;
-  padding: 5px;
+  padding: 7px;
+  border-radius: 5px;
   transition-duration: 0.5s;
 
   &:hover {
